fixed parameter issues with user input when replying tweets with excess and duplicate parameters

+++ main.py
@@ -1,10 +1,10 @@
-import tweepy
 import os
+import time
 import logging
 import logging.handlers as handlers
 from last_seen import get_last_seen_id, set_last_seen_id
 from mention_handler import parse_mention, get_reply_text
-from twitter_api_handler import get_twitter_api, reply_to_tweet
+from twitter_api_handler import get_twitter_api, reply_to_tweet, delete_all_tweets, post_test_tweets
 from mal_api_handler import mal_api_request
 
 rfh = logging.handlers.RotatingFileHandler(
@@ -27,9 +27,16 @@ logging.basicConfig(
 
 
 def main():
+    
+
     log = logging.getLogger()
 
     twitter_api, last_seen_file_name = get_twitter_api()
+
+    #delete test tweets and post them
+    delete_all_tweets(twitter_api)
+    post_test_tweets(twitter_api)
+    #quit()
 
     # Get all mentions tweeted after the last seen tweet id
     last_seen_id = get_last_seen_id(last_seen_file_name)
@@ -56,9 +63,13 @@ def main():
         reply_to_tweet(twitter_api, media_info['image_file_path'], reply_text, mention._json['id'])
 
         # Delete anime cover photo
-        os.remove(media_info['image_file_path'])
+        if media_info['image_file_path'] != '':
+            os.remove(media_info['image_file_path'])
         
         log.info(' ')
+        time.sleep(1)
+        
+    #delete_all_tweets(twitter_api)
 
     # Set the last seen tweet id to the id if the last parsed mention
     last_seen_id = mentions[-1]._json['id']
